Create an CRUD endpoint for managing contacts

+++ src/main/java/com/minh/contacts/repositories/ContactRepository.java
@@ -3,5 +3,5 @@ package com.minh.contacts.repositories;
 import com.minh.contacts.models.Contact;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ContactRepository extends JpaRepository<Contact, Integer> {
+public interface ContactRepository extends JpaRepository<Contact, Long> {
 }
