+++ src/main/java/com/minh/contacts/repositories/SkillRepository.java
@@ -3,5 +3,5 @@ package com.minh.contacts.repositories;
 import com.minh.contacts.models.Skill;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SkillRepository extends JpaRepository<Skill, Integer> {
+public interface SkillRepository extends JpaRepository<Skill, Long> {
 }
